+++ CHANGELOG.md
@@ -4,9 +4,10 @@ Change Log
 ### 12/28/2018 - 1.0.2
 
 #### Added
-* GoogleAds: `Operation` class for all changes.
 * GoogleAds: `AdGroupOperation` with ability to change adgroup bids using `setBid()`
+* GoogleAds: `AdGroupOperation` ability to change name and status of adgroups.
 * GoogleAds: `AdGroupOperation` with a `save()` method that posts changes to Google Ads.
+* GoogleAds: `AdGroupOperation` Added getter methods `getName()`, `getStatus()`, `getCampaignId()`, `getId()`
 
 ### 12/28/2018 - 1.0.1